code change for wxp extensions on Toolkit
