+++ .vscode/launch.json
@@ -5,44 +5,6 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "name": "Debug UI Tests",
-      "type": "node",
-      "request": "launch",
-      "program": "${workspaceFolder}/node_modules/mocha/bin/_mocha",
-      "args": [
-        "-u",
-        "bdd",
-        "--timeout",
-        "999999",
-        "--colors",
-        "${workspaceFolder}/test/end-to-end",
-        "-r",
-        "ts-node/register",
-        "${workspaceFolder}/test/end-to-end/*.ts"
-      ],
-      "internalConsoleOptions": "openOnSessionStart",
-      "runtimeArgs": ["--preserve-symlinks"]
-    },
-    {
-      "name": "Debug Unit Tests",
-      "type": "node",
-      "request": "launch",
-      "program": "${workspaceFolder}/node_modules/mocha/bin/_mocha",
-      "args": [
-        "-u",
-        "bdd",
-        "--timeout",
-        "999999",
-        "--colors",
-        "${workspaceFolder}/test/unit",
-        "-r",
-        "ts-node/register",
-        "${workspaceFolder}/test/unit/*.test.ts"
-      ],
-      "internalConsoleOptions": "openOnSessionStart",
-      "runtimeArgs": ["--preserve-symlinks"]
-    },
-    {
       "name": "Excel Desktop (Edge Chromium)",
       "type": "msedge",
       "request": "attach",
@@ -67,7 +29,6 @@
       "name": "Outlook Desktop (Edge Chromium)",
       "type": "msedge",
       "request": "attach",
-      "useWebView": true,
       "port": 9229,
       "timeout": 600000,
       "webRoot": "${workspaceRoot}",
