+++ manifest.json
@@ -1,0 +1,223 @@
+{
+  "$schema": "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json",
+  "id": "f2b103f1-1ab1-4e1b-8f0b-072aa3d4e19d",
+  "manifestVersion": "devPreview",
+  "version": "1.0.0",
+  "name": {
+    "short": "Contoso Task Pane Add-in",
+    "full": "Contoso Task Pane Add-in"
+  },
+  "description": {
+    "short": "A template to get started.",
+    "full": "This is the template to get started."
+  },
+  "developer": {
+    "name": "Contoso",
+    "websiteUrl": "https://www.contoso.com",
+    "privacyUrl": "https://www.contoso.com/privacy",
+    "termsOfUseUrl": "https://www.contoso.com/servicesagreement"
+  },
+  "icons": {
+    "outline": "assets/outline.png",
+    "color": "assets/color.png"
+  },
+  "accentColor": "#230201",
+  "localizationInfo": {
+    "defaultLanguageTag": "en-us",
+    "additionalLanguages": []
+  },
+  "authorization": {
+    "permissions": {
+      "resourceSpecific": [
+        {
+          "name": "MailboxItem.Read.User",
+          "type": "Delegated"
+        },
+        {
+          "name": "Document.ReadWrite.User",
+          "type": "Delegated"
+        }
+      ]
+    }
+  },
+  "validDomains": ["contoso.com"],
+  "extensions": [
+    {
+      "requirements": {
+        "scopes": ["mail", "workbook", "document", "presentation"],
+        "capabilities": [
+          { "name": "Mailbox", "minVersion": "1.3" },
+          { "name": "AddinCommands", "minVersion": "1.1" }
+        ]
+      },
+      "runtimes": [
+        {
+          "requirements": {
+            "capabilities": [
+              { "name": "Mailbox", "minVersion": "1.3" }
+            ]
+          },
+          "id": "TaskPaneRuntime",
+          "type": "general",
+          "code": {
+            "page": "https://localhost:3000/taskpane.html"
+          },
+          "lifetime": "short",
+          "actions": [
+            {
+              "id": "TaskPaneRuntimeShow",
+              "type":"openPage",
+              "pinnable": false,
+              "view": "dashboard"
+            }
+          ]
+        },
+        {
+          "id": "CommandsRuntime",
+          "type": "general",
+          "code": {
+            "page": "https://localhost:3000/commands.html",
+            "script": "https://localhost:3000/commands.js"
+          },
+          "lifetime": "short",
+          "actions": [
+            {
+              "id": "action",
+              "type": "executeFunction",
+              "displayName": "action"
+            }
+          ]
+        },
+        {
+          "requirements": {
+            "capabilities": [
+              { "name": "AddinCommands", "minVersion": "1.1" }
+            ]
+          },
+          "id": "TaskPaneRuntime",
+          "type": "general",
+          "code": {
+            "page": "https://localhost:3000/taskpane.html"
+          },
+          "lifetime": "short",
+          "actions": [
+            {
+              "id": "TaskPaneRuntimeShow",
+              "type":"openPage",
+              "pinnable": false,
+              "view": "dashboard"
+            }
+          ]
+        }
+      ],
+      "ribbons": [
+        {
+          "contexts": [
+            "mailRead"
+          ],
+          "tabs": [
+            {
+              "builtInTabId": "TabDefault",
+              "groups": [
+                {
+                  "id": "msgReadGroup",
+                  "label": "Contoso Add-in",
+                  "icons": [
+                    { "size": 16, "file": "https://localhost:3000/assets/icon-16.png" },
+                    { "size": 32, "file": "https://localhost:3000/assets/icon-32.png" },
+                    { "size": 80, "file": "https://localhost:3000/assets/icon-80.png" }
+                  ],
+                  "controls": [
+                    {
+                      "id": "msgReadOpenPaneButton",
+                      "type": "button",
+                      "label": "Show Taskpane",
+                      "icons": [
+                        { "size": 16, "file": "https://localhost:3000/assets/icon-16.png" },
+                        { "size": 32, "file": "https://localhost:3000/assets/icon-32.png" },
+                        { "size": 80, "file": "https://localhost:3000/assets/icon-80.png" }
+                      ],
+                      "supertip": {
+                        "title": "Show Taskpane",
+                        "description": "Opens a pane displaying all available properties."
+                      },
+                      "actionId": "TaskPaneRuntimeShow"
+                    },
+                    {
+                      "id": "ActionButton",
+                      "type": "button",
+                      "label": "Perform an action",
+                      "icons": [
+                        { "size": 16, "file": "https://localhost:3000/assets/icon-16.png" },
+                        { "size": 32, "file": "https://localhost:3000/assets/icon-32.png" },
+                        { "size": 80, "file": "https://localhost:3000/assets/icon-80.png" }
+                      ],
+                      "supertip": {
+                        "title": "Perform an action",
+                        "description": "Perform an action when clicked."
+                      },
+                      "actionId": "action"
+                    }
+                  ]
+                }
+              ]
+            }
+          ]
+        },
+        {
+          "contexts": [
+            "default"
+          ],
+          "tabs": [
+            {
+              "builtInTabId": "TabHome",
+              "groups": [
+                {
+                  "id": "msgReadGroup",
+                  "label": "Contoso Add-in",
+                  "icons": [
+                    { "size": 16, "file": "https://localhost:3000/assets/icon-16.png" },
+                    { "size": 32, "file": "https://localhost:3000/assets/icon-32.png" },
+                    { "size": 80, "file": "https://localhost:3000/assets/icon-80.png" }
+                  ],
+                  "controls": [
+                    {
+                      "id": "msgReadOpenPaneButton",
+                      "type": "button",
+                      "label": "Hello Excel",
+                      "icons": [
+                        { "size": 16, "file": "https://localhost:3000/assets/icon-16.png" },
+                        { "size": 32, "file": "https://localhost:3000/assets/icon-32.png" },
+                        { "size": 80, "file": "https://localhost:3000/assets/icon-80.png" }
+                      ],
+                      "supertip": {
+                        "title": "Show Taskpane",
+                        "description": "Opens a pane displaying all available properties."
+                      },
+                      "actionId": "TaskPaneRuntimeShow"
+                    },
+                    {
+                      "id": "ActionButton",
+                      "type": "button",
+                      "label": "Perform an action",
+                      "icons": [
+                        { "size": 16, "file": "https://localhost:3000/assets/icon-16.png" },
+                        { "size": 32, "file": "https://localhost:3000/assets/icon-32.png" },
+                        { "size": 80, "file": "https://localhost:3000/assets/icon-80.png" }
+                      ],
+                      "supertip": {
+                        "title": "Perform an action",
+                        "description": "Perform an action when clicked."
+                      },
+                      "actionId": "action"
+                    }
+                  ]
+                }
+              ]
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
