+++ package.json
@@ -23,15 +23,9 @@
     "lint": "office-addin-lint check",
     "lint:fix": "office-addin-lint fix",
     "prettier": "office-addin-lint prettier",
-    "start": "office-addin-debugging start manifest.xml",
-    "start:desktop": "office-addin-debugging start manifest.xml desktop",
-    "start:desktop:outlook": "office-addin-debugging start manifest.outlook.xml desktop --app outlook",
-    "start:web": "office-addin-debugging start manifest.xml web",
-    "stop": "office-addin-debugging stop manifest.xml",
-    "test": "npm run test:unit && npm run test:e2e",
-    "test:e2e": "mocha -r ts-node/register test/end-to-end/*.ts",
-    "test:unit": "mocha -r ts-node/register test/unit/*.test.ts",
-    "validate": "office-addin-manifest validate manifest.xml",
+    "start": "office-addin-debugging start manifest.json",
+    "stop": "office-addin-debugging stop manifest.json",
+    "validate": "office-addin-manifest validate manifest.json",
     "watch": "webpack --mode development --watch"
   },
   "dependencies": {
