--- src/taskpane/word-office-document.js
@@ -1,16 +1,0 @@
-/* global Word console */
-
-const insertText = async (text) => {
-  // Write text to the document.
-  try {
-    await Word.run(async (context) => {
-      let body = context.document.body;
-      body.insertParagraph(text, Word.InsertLocation.end);
-      await context.sync();
-    });
-  } catch (error) {
-    console.log("Error: " + error);
-  }
-};
-
-export default insertText;
